% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{adj_g2_test}
\alias{adj_g2_test}
\title{Two-sided likelihood ratio test for calibration}
\usage{
adj_g2_test(p0, data, alpha)
}
\arguments{
\item{p0}{vector, the category probabilities under the null hypothesis}

\item{data}{vector of observed counts in each category.}

\item{alpha}{the desired significance level}
}
\value{
a list
$pval The calculated p-value
$result A boolean, TRUE if the nul is rejected, FALSE if it is not.
}
\description{
A two-sided likelihood ratio test for model calibration. Similar in theory to the Bayes factor test, this
tests the hypothesis that \eqn{H_0: p=p_0} against \eqn{H_1:p\ne p_0}. We present the G^2/q' statistic presented in Smith et al 1981 given by

\eqn{
G^2=\frac{2}{q'}\sum_{i=1}^{d+1}n_i\log(n_i/(np_{0,i}))
}
where

\eqn{q'=1+\frac{1}{6dn}\left(\sum_i \frac{1}{p_{0,i}}-1+\frac{1}{n}\sum_i\left(\frac{1}{p_{0,i}}-\frac{1}{p_{0,i}^2}\right)\right)}

which has an asymptotic \eqn{\chi^2} distribution with d degrees of freedom in the case where there are d+1 categories.
}
