% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_test.R
\name{rel_plot_spring}
\alias{rel_plot_spring}
\title{Reliability plot estimation}
\usage{
rel_plot_spring(epsvals, ground_truth, preds, predstds, ptest)
}
\arguments{
\item{epsvals}{The spring nonlinearity parameters}

\item{ground_truth}{The true state (position OR velocity)}

\item{preds}{A time series of predicted positions OR velocities}

\item{predstds}{The predictive standard deviations}

\item{ptest}{A vector of the nominal coverage levels.}
}
\value{
Empirical coverage estimates
}
\description{
Estimates coverage rates and generates reliability plots for the
probabilistic model based on inputs. Currently under development, so just returns the
empirical coverages
}
