% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_test.R
\name{sim_spring}
\alias{sim_spring}
\title{Simulation script for a nonlinear spring}
\usage{
sim_spring(
  nsteps = 100,
  dt = 0.01,
  x0 = c(1, 0),
  k = 2,
  m = 1,
  g = 0.3,
  epsilon = 0,
  sigma_w = 0.001
)
}
\arguments{
\item{nsteps}{Scalar, the number of time steps to run the system for. Default 100}

\item{dt}{time step. Default 0.01}

\item{x0}{array, vector, or similar. The initial position and velocity of the spring. Default c(1,0)}

\item{k}{Scalar stiffness parameter of the spring. Default 2.}

\item{m}{Scalar, mass of the object. Default 1}

\item{g}{Scalar damping parameter. Default 0.3}

\item{epsilon}{Scalar nonlinearity term. Default 0(linear spring)}

\item{sigma_w}{Scalar, standard deviation of the random noise term added to the velocity. Default \eqn{10^-3}.}
}
\value{
The state vector (position and velocity) of the mass.
}
\description{
Uses the 4th order Runge-Kutta method to solve the differential
equation
\eqn{mx''+gx'+kx+\epsilon x^3=0}
corresponding to a nonlinear spring-mass-damper system with mass m. This is also known
as the unforced Duffing oscillator. The model solves the differential equation subject to random forcing on the
velocity and outputs the location and velocity of the mass as a function of time.
}
