% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{t_trunc_test}
\alias{t_trunc_test}
\title{Truncated-T test for the calibration}
\usage{
t_trunc_test(p0, data, alpha)
}
\arguments{
\item{p0}{vector, the category probabilities under the null hypothesis}

\item{data}{vector of observed counts in each category.}

\item{alpha}{the desired significance level}
}
\value{
a list
$cv The critical value
$ttrunc The value of \eqn{T_{trunc}}
$result A boolean, TRUE if the nul is rejected, FALSE if it is not.
}
\description{
A two-sided calibration test for model calibration based on the test derived in Balakrishnan and Wasserman 2019.
If there are d+1 categories, which corresponds to d nominal coverage levels, the test statistic \eqn{T_{trunc}} is given by
\eqn{
T_{trunc}=\sum_i\frac{(n_i-np_{0,i})^2-n_i}{t_i}
}
where
\eqn{t_i=max(1/(d+1), p_{0,i})}.
The null hypothesis is rejected at the alpha level if
\eqn{
T_{trunc}>n\sqrt{\frac{2}{\alpha}\sum_i\left(\frac{p_{0,i}}{t_i}\right)^2}
}
}
