% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{two_sided_bayes_factor}
\alias{two_sided_bayes_factor}
\title{Two-sided Bayes factor test for calibration}
\usage{
two_sided_bayes_factor(
  p0,
  data,
  dir_prior = array(0.5, dim = c(1, length(p0)))
)
}
\arguments{
\item{p0}{vector of null category probabilities.}

\item{data}{vector of the same length as p0 containing the number of observed members in each category.}

\item{dir_prior}{prior parameters for the dirichlet distribution. Default is all \eqn{\gamma_i=1/2} which is the Jeffreys prior.}
}
\value{
a list:
$logfactor The natural log of the two-sided Bayes factor BF_{10}
$interpretation The interpretation of the Bayes factor as described in Kass and Raftery 1995
}
\description{
Calculates the Bayes factor \eqn{BF_{10}} for the alternative and null models
for calibration testing. The null model is that the vector \eqn{\vec{p_0}} describes the
true gap probabilities, and the alternative is that it does not. Letting X be the data, n be the total observations,
and m_1 and m_0 be the alternative and null models, the Bayes factor is
\eqn{
BF_{10} = \frac{m_1(X)}{m_0(X)}=\frac{\Gamma(\sum_i\gamma_i)}{\Gamma(n+\sum_i\gamma_i)}\frac{\prod_i \Gamma(n_i+\gamma_i)/\Gamma(\gamma_i)}{(p_{0,i})^n_i}
}
where \eqn{\vec{\gamma}} is the vector of parameters for the Dirichlet prior distribution. This is an analogue of the frequentist likelihood ratio test, and has many interpretations.
See Kass and Raftery 1995 for one potential interpretation.
}
