% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_test.R
\name{kalman_filter_spring}
\alias{kalman_filter_spring}
\title{Kalman filter for the spring-mass damper system assuming perfect linearity.}
\usage{
kalman_filter_spring(
  truth,
  P0 = diag(c(0.01, 0.01)),
  Q = diag(c(0, 0.001^2)),
  R = diag(c(0.001, 0.001)),
  k = 2,
  m = 1,
  g = 0.3,
  dt = 0.01
)
}
\arguments{
\item{truth}{2\times nsteps array containing the mass position along the first row and velocity along the second.}

\item{Q}{A covariance matrix for the perturbations applied to the spring-mass system. Default is the degenerate covariance matrix
[\link{0,0};\link{0,0.001^2}] corresponding to no perturbations to position and Gaussian perturbations to velocity with mean 0 and sstandard deviation 0.001.}

\item{R}{2\times 2 observation error covariance. Default [\link{0.001,0};\link{0,0.001}]. This represents the observation errors through which the true state is
observed. If you wish to simulate a system where velocity is unobserved, set the second diagonal entry to a large value, i.e. 1e6.}

\item{k, m, g}{The spring stiffness, mass, and damping. Defaults are 2,1, 0.3 respectively.}

\item{dt}{The time step. Default is 0.01.}

\item{p0}{2\times 2 matrix. The initial guess for the covariance matrix of the predictions. Default is \eqn{P0=[[0.01,0];[0,0.01]]}.}
}
\value{
a list:
$filt: The Kalman filtered state vector
$pred: The predicted state, used for the calibration estimation.
$predcov: the predictive covariance matrices.
$postcov: Posterior covariance matrices
$truth: The true state, same as the corresponding input parameter
}
\description{
Using the 4th order Runge-Kutta method on the ODE
\eqn{mx''+gx'+kx=0}, performs Kalman filtering of the position and velocity of the spring
mass damper system under observation error covariance matrix R. The method assumes that the spring is
linear, introducing model form error that prevents accurate model calibration. This is implemented as a
demonstration of the package's calibration checking methodology.
}
